% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utilities.R
\name{array2vec}
\alias{array2vec}
\title{Convert array to binary vector in row major order}
\usage{
array2vec(arr.mat)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Convert array to binary vector in row major order
}
