% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utilities.R
\name{check.analysis}
\alias{check.analysis}
\title{Check a Q feathering pattern against a K
Check a Q feathering pattern against a K. Handy for checking miss-IDed fps}
\usage{
check.analysis(
  a.Q.img.num,
  a.K.set.num,
  K.set.fnames,
  an.fp.info.mat,
  fp.file.names,
  reg.fill.typ = "255",
  path.to.fp.images,
  path.to.fp.fits,
  plot.level = 0
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Check a Q feathering pattern against a K
Check a Q feathering pattern against a K. Handy for checking miss-IDed fps
}
