% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_utilities.R
\name{chop.and.fill}
\alias{chop.and.fill}
\title{Utility function for XXXXXX}
\usage{
chop.and.fill(
  tmpl.in.ref.padded,
  reference.img,
  fill.typ,
  h.pad.leng,
  w.pad.leng
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function depends on the Rvision library which must be installed
}
