% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_utilities.R
\name{emp.pix.dropout.rates}
\alias{emp.pix.dropout.rates}
\title{Utility function for XXXXXX}
\usage{
emp.pix.dropout.rates(img.list)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function depends on the Rvision library which must be installed
}
