% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_utilities.R
\name{geom_analysis}
\alias{geom_analysis}
\title{Extract some geometric parameters of a thresholded feathering pattern
XXXX}
\usage{
geom_analysis(
  thresh.img.obj,
  black.pixel.code = 0,
  plotQ = F,
  plot.img.type = "image.matrix"
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Extract some geometric parameters of a thresholded feathering pattern
XXXX
}
