% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utilities.R
\name{image.list2vec.mat}
\alias{image.list2vec.mat}
\title{Convert a list of images to a matrix of row vectors}
\usage{
image.list2vec.mat(img.list)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Convert a list of images (i.e. an image stack) to a matrix of row vectors
}
\details{
Assumes all images in the stack are of the same dimension
}
