% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utilities.R
\name{ind2sub}
\alias{ind2sub}
\title{Vector index to matrix subscript}
\usage{
ind2sub(idx, num.row, col.majQ = F)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Vector index to matrix subscript
}
