% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utilities.R
\name{match.analysis}
\alias{match.analysis}
\title{Analyze probs of feathing patters on each feathering pattern set
XXXX}
\usage{
match.analysis(
  an.idx,
  prob.vec,
  an.fp.info.mat,
  an.lbl.names,
  min.prob = 0.001,
  printQ = F,
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Analyze probs of feathing patters on each feathering pattern set
XXXX
}
