% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_utilities.R
\name{mean.stack.img}
\alias{mean.stack.img}
\title{Mean image of a stack}
\usage{
\method{mean}{stack.img}(img.list, scale.factor = 255)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Mean image of a stack, thresholded by the scale.factor
}
\details{
The function depends on the Rvision library which must be installed
}
