% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utilities.R
\name{mle.fit.stack}
\alias{mle.fit.stack}
\title{Wrapper to quick fit a stack of registered images}
\usage{
mle.fit.stack(img.list, num.sims, num.iters = 1, grad.tol = 0.001)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Wrapper to quick fit a stack of registered images. If this goes wrong, fit the stack step by
step to see where it fails.
}
