% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utilities.R
\name{node.num2idxs}
\alias{node.num2idxs}
\title{Convert node number to matrix indices in a rectangular array of nodes}
\usage{
node.num2idxs(idx, num.row)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Convert node number to matrix indices in a rectangular array of nodes
}
\details{
Re-name of ind2sub and removed cmQ option for speed
}
