% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{plot.overlay}
\alias{plot.overlay}
\title{XXXXXX}
\usage{
\method{plot}{overlay}(
  template.img,
  reference.img,
  h.min.overlap = 1,
  w.min.overlap = 1,
  x.max,
  y.max,
  ref.lo = 0,
  ref.hi = 2,
  templ.lo = 4,
  templ.hi = 6,
  delay = 0.1
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function depends on the Rvision library which must be installed
}
