% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{plot.stack}
\alias{plot.stack}
\title{Plot an image stack}
\usage{
\method{plot}{stack}(
  img.list,
  num.cycles = 1,
  type = "Rvision",
  main.title = NULL,
  delay = 0.5
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Plot an image stack by plotting one image after the next
}
\details{
The function depends on the Rvision library which must be installed
}
