% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utilities.R
\name{register.template}
\alias{register.template}
\title{Register (best align) a template to a reference.}
\usage{
register.template(
  template.img,
  reference.img,
  h.min.overlap = 1,
  w.min.overlap = 1,
  method = "SQDIFF",
  fill.typ = "opposite",
  printQ = F,
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function depends on the Rvision library which must be installed
}
