% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_utilities.R
\name{simulate.pix.pattern}
\alias{simulate.pix.pattern}
\title{Simulate pixel patterns}
\usage{
\method{simulate}{pix.pattern}(img.list, nsims = 1)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Given an input set of pixel patterns, simulate new patterns based on the input pixel
patterns dropout rates.
}
