% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utilities.R
\name{vec2array}
\alias{vec2array}
\title{Convert binary vector to an array assuming row major order}
\usage{
vec2array(arr.vec, num.rows, num.cols, white.pix = 1, black.pix = 0)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Convert binary vector to an array assuming row major order
}
