% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_utilities.R
\name{emp.pix.dropout.rates}
\alias{emp.pix.dropout.rates}
\title{Utility function for computing pixel drop out rates in an array of black and white pixels}
\usage{
emp.pix.dropout.rates(img.list)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The pixel drop out rate is the probability the pixel will be white (255) given that it was
black (0) at least once in a stack of registered images
}
