% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_utilities.R
\name{thresh.img.obj}
\alias{thresh.img.obj}
\title{XXXXXX}
\usage{
thresh.img.obj(img.obj, scale.factor = 255)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function depends on the Rvision library which must be installed
}
